+++ src/main/java/com/Rest/GolfMax/API/Controllers/LoginController.java
@@ -1,0 +1,25 @@
+package com.Rest.GolfMax.API.Controllers;
+
+import com.Rest.GolfMax.API.Models.User;
+import com.Rest.GolfMax.API.Services.UserService;
+import org.jetbrains.annotations.NotNull;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequestMapping("/users")
+public class LoginController {
+    @Autowired
+    private UserService userService;
+
+    @PostMapping("/login")
+    public ResponseEntity<User> login(@RequestBody @NotNull User user) {
+        User storedUser = userService.getStoredUserData(user.getUsername(), user.getPassword());
+        if (storedUser == null)
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        else
+            return new ResponseEntity<>(storedUser, HttpStatus.OK);
+    }
+}
